+++ changelog.md
@@ -1,3 +1,9 @@
+# 1.0.2
+
+- Added `addValueChangeListener` and `removeValueChangeListener` in `GM.` and `GM_` APIs
+- Improved type generics for `setValue`, `setValues`, `getValue`, `getValues` in `GM.` and `GM_` APIs
+- Fixed typo in Violentmonkey name
+
 # 1.0.1
 
 - Fixed README